Changes related to File Type and File Size
