+++ app/lib/common/utils/utils.dart
@@ -165,6 +165,34 @@ Future<bool> openLink(String target, BuildContext context) async {
   }
 }
 
+String getFileSize(int bytes) {
+  if (bytes <= 0) return '0 B';
+  const suffixes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+  var i = (log(bytes) / log(1024)).floor();
+  return '${(bytes / pow(1024, i)).toStringAsFixed(1)} ${suffixes[i]}';
+}
+
+String documentTypeFromFileExtension(String fileExtension) {
+  switch (fileExtension) {
+    case 'png':
+    case 'jpg':
+    case 'jpeg':
+      return 'Image';
+    case 'mov':
+    case 'mp4':
+      return 'Video';
+    case 'mp3':
+    case 'wav':
+      return 'Audio';
+    case 'pdf':
+      return 'PDF';
+    case 'txt':
+      return 'Text File';
+    default:
+      return '';
+  }
+}
+
 Future<void> shareTextToWhatsApp(
   BuildContext context, {
   required String text,
