+++ app/lib/features/pins/pages/create_pin.dart
@@ -4,13 +4,13 @@ import 'package:acter/common/models/types.dart';
 import 'package:acter/common/toolkit/buttons/inline_text_button.dart';
 import 'package:acter/common/toolkit/buttons/primary_action_button.dart';
 import 'package:acter/common/utils/routes.dart';
+import 'package:acter/common/utils/utils.dart';
 import 'package:acter/common/widgets/edit_title_sheet.dart';
 import 'package:acter/common/widgets/input_text_field.dart';
 import 'package:acter/common/widgets/render_html.dart';
 import 'package:acter/common/widgets/spaces/select_space_form_field.dart';
 import 'package:acter/features/attachments/actions/handle_selected_attachments.dart';
 import 'package:acter/features/pins/actions/attachment_leading_icon.dart';
-import 'package:acter/features/pins/actions/select_pin_attachments.dart';
 import 'package:acter/features/pins/actions/set_pin_description.dart';
 import 'package:acter/features/pins/actions/set_pin_links.dart';
 import 'package:acter/features/pins/models/create_pin_state/create_pin_state.dart';
